Add prepared statement caching to pgsql and sqlite queues

+++ src/Phive/Queue/Db/PDO/PgSqlPDOQueue.php
@@ -4,6 +4,11 @@ namespace Phive\Queue\Db\PDO;
 
 class PgSqlPDOQueue extends AbstractPDOQueue
 {
+    /**
+     * @var \PDOStatement
+     */
+    protected $popStatement;
+
     public function __construct(\PDO $conn, $tableName)
     {
         if ('pgsql' != $conn->getAttribute(\PDO::ATTR_DRIVER_NAME)) {
@@ -19,17 +24,20 @@ class PgSqlPDOQueue extends AbstractPDOQueue
      */
     public function pop()
     {
-        $sql = 'SELECT id FROM '.$this->tableName.' WHERE eta <= :eta ORDER BY eta, id LIMIT 1';
-        $sql = 'DELETE FROM '.$this->tableName.' WHERE id = ('.$sql.') RETURNING item';
+        if (!$this->popStatement) {
+            $sql = 'SELECT id FROM '.$this->tableName.' WHERE eta <= :eta ORDER BY eta, id LIMIT 1';
+            $sql = 'DELETE FROM '.$this->tableName.' WHERE id = ('.$sql.') RETURNING item';
 
-        $stmt = $this->prepareStatement($sql);
-        $stmt->bindValue(':eta', time(), \PDO::PARAM_INT);
+            $this->popStatement = $this->prepareStatement($sql);
+        }
+
+        $this->popStatement->bindValue(':eta', time(), \PDO::PARAM_INT);
 
         $this->conn->beginTransaction();
 
         try {
             $this->execute('LOCK TABLE '.$this->tableName.' IN EXCLUSIVE MODE');
-            $this->executeStatement($stmt);
+            $this->executeStatement($this->popStatement);
 
             $this->conn->commit();
         } catch (\Exception $e) {
@@ -37,6 +45,6 @@ class PgSqlPDOQueue extends AbstractPDOQueue
             throw $e;
         }
 
-        return $stmt->fetchColumn();
+        return $this->popStatement->fetchColumn();
     }
 }
